pyproject.toml
[tool.poetry]
name = "autogen-tutorial"
version = "0.1.0"
description = ""
authors = ["Yuki Oshima <39944763+os1ma@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
python-dotenv = "^1.0.1"
pyautogen = "^0.3.0"
autogenstudio = "^0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
